move completion test, LSAndTsDocResovler to new file
