+++ packages/language-server/src/plugins/typescript/TypeScriptPlugin.ts
@@ -26,7 +26,6 @@ import {
 } from '../../lib/documents';
 import { LSConfigManager, LSTypescriptConfig } from '../../ls-config';
 import { pathToUrl } from '../../utils';
-import { getLanguageServiceForDocument } from './service';
 import {
     convertRange,
     getScriptKindFromAttributes,
@@ -35,7 +34,6 @@ import {
     findTsConfigPath,
     getScriptKindFromFileName,
 } from './utils';
-import { TypescriptDocument } from './TypescriptDocument';
 import {
     CodeActionsProvider,
     DefinitionsProvider,
@@ -52,6 +50,7 @@ import {
 import { SnapshotManager } from './SnapshotManager';
 import { DocumentSnapshot, INITIAL_VERSION } from './DocumentSnapshot';
 import { CompletionEntryWithIdentifer, CompletionsProviderImpl } from './features/CompletionProvider';
+import { LSAndTSDocResovler } from './LSAndTSDocResovler';
 
 export class TypeScriptPlugin
     implements
@@ -348,34 +347,3 @@ export class TypeScriptPlugin
     }
 }
 
-export class LSAndTSDocResovler {
-    constructor(
-        private readonly docManager: DocumentManager
-    ) { }
-
-    createDocument = (fileName: string, content: string) => {
-        const uri = pathToUrl(fileName);
-        const document = this.docManager.openDocument({
-            languageId: '',
-            text: content,
-            uri,
-            version: 0,
-        });
-        this.docManager.lockDocument(uri);
-        return new TypescriptDocument(document);
-    };
-
-    private documents = new Map<Document, TypescriptDocument>();
-
-    public getLSAndTSDoc(document: Document) {
-        let tsDoc = this.documents.get(document);
-        if (!tsDoc) {
-            tsDoc = new TypescriptDocument(document);
-            this.documents.set(document, tsDoc);
-        }
-
-        const lang = getLanguageServiceForDocument(tsDoc, this.createDocument);
-
-        return { tsDoc, lang };
-    }
-}
