+++ packages/language-server/test/plugins/typescript/TypescriptPlugin.test.ts
@@ -1,11 +1,8 @@
 import * as assert from 'assert';
 import * as path from 'path';
 import { dirname, join } from 'path';
-import { mkdirSync, rmdirSync } from 'fs';
 import ts from 'typescript';
 import {
-    CompletionItem,
-    CompletionItemKind,
     FileChangeType,
     Hover,
     Position,
@@ -107,117 +104,6 @@ describe('TypescriptPlugin', () => {
                 name: 'bla',
             },
         ]);
-    });
-
-    it('provides completions', async () => {
-        const { plugin, document } = setup('completions.svelte');
-
-        const completions = plugin.getCompletions(document, Position.create(0, 49), '.');
-
-        assert.ok(
-            Array.isArray(completions && completions.items),
-            'Expected completion items to be an array',
-        );
-        assert.ok(completions!.items.length > 0, 'Expected completions to have length');
-
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const { data, ...withoutData } = completions!.items[0];
-
-        assert.deepStrictEqual(withoutData, <CompletionItem>{
-            label: 'b',
-            insertText: undefined,
-            kind: CompletionItemKind.Method,
-            sortText: '0',
-            commitCharacters: ['.', ',', '('],
-            preselect: undefined,
-        });
-    });
-
-    it('provides completion resolve info', async () => {
-        const { plugin, document } = setup('completions.svelte');
-
-        const completions = plugin.getCompletions(document, Position.create(0, 49), '.');
-
-        const { data } = completions!.items[0];
-
-        // uri would not be the same, so only check if exist;
-        assert.notEqual(data, null);
-        const { uri, ...withoutUri } = data!;
-        assert.ok(uri);
-
-        assert.deepStrictEqual(withoutUri, {
-            hasAction: undefined,
-            insertText: undefined,
-            isRecommended: undefined,
-            kind: 'method',
-            kindModifiers: '',
-            name: 'b',
-            position: {
-                character: 49,
-                line: 0
-            },
-            replacementSpan: undefined,
-            sortText: '0',
-            source: undefined,
-        });
-    });
-
-    it('resolve completion and provide documentation', async () => {
-        const { plugin, document } = setup('documentation.svelte');
-
-        const completions = plugin.getCompletions(document, Position.create(4, 8), '(');
-
-        const { documentation, detail } = await plugin.resolveCompletion(
-            document,
-            completions?.items[0]!
-        );
-
-        assert.deepStrictEqual(detail, '(alias) function foo(): boolean\nimport foo');
-        assert.deepStrictEqual(documentation, 'bars');
-    });
-
-    it('provides import completions for directory', async () => {
-        const { plugin, document } = setup('importcompletions.svelte');
-        const mockDirName = 'foo';
-        const mockDirPath = path.join(__dirname, 'testfiles', mockDirName);
-
-        mkdirSync(mockDirPath);
-
-        try {
-            const completions = plugin.getCompletions(document, Position.create(0, 27), '/');
-            const mockedDirImportCompletion = completions?.items
-                .find(item => item.label === mockDirName);
-
-            assert.notEqual(
-                mockedDirImportCompletion,
-                undefined,
-                `can't provides completions on directory`
-            );
-            assert.equal(mockedDirImportCompletion?.kind, CompletionItemKind.Folder);
-        } finally {
-            rmdirSync(mockDirPath);
-        }
-    });
-
-    it('resolve auto import completion', async () => {
-        const { plugin, document } = setup('importcompletions.svelte');
-
-        const completions = plugin.getCompletions(document, Position.create(0, 40));
-        document.version++;
-
-        const item = completions?.items.find(item => item.label === 'blubb');
-
-        assert.equal(item?.additionalTextEdits, undefined);
-        assert.equal(item?.detail, undefined);
-
-        const { additionalTextEdits, detail } = await plugin.resolveCompletion(document, item!);
-
-        assert.strictEqual(detail, 'Auto import from ./definitions\nfunction blubb(): boolean');
-
-        assert.strictEqual(
-            additionalTextEdits![0]?.newText.replace('\r', '').replace('\n', ''),
-            `import { blubb } from './definitions'`,
-        );
     });
 
     it('provides definitions within svelte doc', () => {
